requirements.txt
camel-ai
playwright
python-dotenv
pyyaml
aiohttp
groq
numpy
pillow
requests-oauthlib
requests
unstructured
pandas
cryoet_data_portal
